feat: introduce unified interface for creating of separate db instances
