+++ btc_indexer/apps/main/Cargo.toml
@@ -18,6 +18,6 @@ btc_indexer_server = { workspace = true }
 config_parser = { workspace = true }
 dotenv = { workspace = true }
 global_utils = { workspace = true }
-persistent_storage = { workspace = true }
+local_db_store_indexer = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
