+++ btc_indexer/apps/main/src/main.rs
@@ -3,7 +3,7 @@ use std::{net::IpAddr, str::FromStr};
 use btc_indexer_internals::indexer::{BtcIndexer, IndexerParams};
 use config_parser::config::{BtcRpcCredentials, ConfigVariant, ServerConfig};
 use global_utils::{env_parser::lookup_ip_addr, logger::init_logger};
-use persistent_storage::{config::PostgresDbCredentials, init::PostgresRepo};
+use local_db_store_indexer::{PostgresDbCredentials, init::LocalDbIndexer};
 use tokio::net::TcpListener;
 
 #[tokio::main]
@@ -11,8 +11,8 @@ async fn main() -> anyhow::Result<()> {
     let _logger_guard = init_logger();
     let app_config = ServerConfig::init_config(ConfigVariant::Production)?;
     // let app_config = ServerConfig::init_config(ConfigVariant::init())?;
-    let (btc_creds, postgres_creds) = (BtcRpcCredentials::new()?, PostgresDbCredentials::new()?);
-    let db_pool = PostgresRepo::from_config(postgres_creds).await?.into_shared();
+    let (btc_creds, postgres_creds) = (BtcRpcCredentials::new()?, PostgresDbCredentials::from_envs()?);
+    let db_pool = LocalDbIndexer::from_config(postgres_creds).await?;
     let btc_indexer = BtcIndexer::with_api(IndexerParams {
         btc_rpc_creds: btc_creds,
         db_pool: db_pool.clone(),
