+++ btc_indexer/crates/indexer_internals/Cargo.toml
@@ -12,7 +12,6 @@ bitcoincore-rpc = { workspace = true }
 config_parser = { workspace = true }
 global_utils = { workspace = true }
 ordinals = { workspace = true }
-persistent_storage = { workspace = true }
 serde = { workspace = true }
 sqlx = { workspace = true }
 thiserror = { workspace = true }
@@ -22,6 +21,7 @@ tokio = { workspace = true }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
+local_db_store_indexer = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
