+++ btc_indexer/crates/local_db_store/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "local_db_store_indexer"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+readme.workspace = true
+
+[dependencies]
+persistent_storage = { workspace = true }
+sqlx = { workspace = true }
+anyhow = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+serde_json = { workspace = true }
+async-trait = { workspace = true }
+global_utils = { workspace = true }
+serde = { workspace = true }
+uuid = { workspace = true }
+url = { workspace = true }
+
+[dev-dependencies]
+bitcoin = { workspace = true }
+dotenv = { workspace = true }
+tokio = { workspace = true }
