+++ btc_indexer/crates/local_db_store/fixtures/btc_indexer_work_checkpoints.sql
