+++ btc_indexer/crates/local_db_store/fixtures/user_request_stats.sql
