+++ btc_indexer/crates/local_db_store/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod error;
+pub mod init;
+pub mod schemas;
+pub use persistent_storage::{config::*, init::PersistentRepoTrait};
