+++ btc_indexer/crates/local_db_store/src/schemas.rs
