+++ btc_indexer/crates/local_db_store/tests/utils.rs
