+++ crates/persistent_storage/example.env
@@ -1,13 +1,6 @@
 APP_ENVIRONMENT="local" # "local" or "production"
 RUST_LOG=trace
 
-POSTGRES_HOST_TESTING="localhost"
-POSTGRES_USER_TESTING="default"
-POSTGRES_PASSWORD_TESTING="12345678"
-POSTGRES_PORT_TESTING="5432"
-POSTGRES_DB_TESTING=testing_db_name
-DATABASE_URL_TESTING="postgres://${POSTGRES_USER_TESTING}:${POSTGRES_PASSWORD_TESTING}@${POSTGRES_HOST_TESTING}:${POSTGRES_PORT_TESTING}/${POSTGRES_DB_TESTING}"
-
 POSTGRES_USER=admin_manager
 POSTGRES_PASSWORD=password
 POSTGRES_HOST=localhost
