+++ crates/persistent_storage/src/error.rs
@@ -3,7 +3,7 @@ use sqlx::migrate::MigrateError;
 
 #[derive(Debug, thiserror::Error)]
 pub enum DbInitError {
-    #[error("Unable to retreive env variable for db initialization, error: {0}")]
+    #[error("Unable to retrieve env variable for db initialization, error: {0}")]
     UnableToRetrieveEnvVar(String),
     #[error("Failed to establish connection with db, please check url [err: {0}, url: {1}]")]
     FailedToEstablishDbConn(sqlx::Error, String),
@@ -14,11 +14,7 @@ pub enum DbInitError {
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum DbError {
-    #[error("Failed to open Pg client, error: {0}")]
-    DbError(#[from] sqlx::Error),
-    #[error("Failed to convert type from json, error: {0}")]
-    SerdeJsonError(#[from] serde_json::error::Error),
+pub enum DbGetConnError {
+    #[error("Unable to retrieve db connection, error: {0}")]
+    UnableToRetrieveConnection(#[from] sqlx::Error),
 }
-
-pub type Result<T> = std::result::Result<T, DbError>;
