+++ crates/persistent_storage/src/init.rs
@@ -5,8 +5,8 @@ use sqlx::{PgPool, Pool, Postgres, pool::PoolConnection};
 use tracing::{instrument, trace};
 
 use crate::{
-    config::{PostgresDbCredentials, PostgresDbTestingCredentials},
-    error::{DbError, DbInitError},
+    config::PostgresDbCredentials,
+    error::{DbGetConnError, DbInitError},
 };
 
 pub type PostgresPool = Pool<Postgres>;
@@ -18,54 +18,23 @@ pub struct PostgresRepo {
     pub pool: PostgresPool,
 }
 
-#[derive(Debug, Clone)]
-pub struct PostgresRepoTesting {
-    pub pool: PostgresPool,
-}
-
 /// Trait for implementing Persistent storage that'd use Postgres
 #[async_trait]
 pub trait PersistentRepoTrait: Send + Sync {
-    async fn get_conn(&self) -> crate::error::Result<PersistentDbConn>;
-}
-
-impl PostgresRepoTesting {
-    #[instrument(level = "trace", ret)]
-    pub async fn from_config(creds: PostgresDbTestingCredentials) -> Result<Self, DbInitError> {
-        trace!("Creating PG(testing) connection pool with creds: {:?}", creds);
-        let pool = PgPool::connect(&creds.url)
-            .await
-            .map_err(|x| DbInitError::FailedToEstablishDbConn(x, creds.url.clone()))?;
-        trace!(db_url = creds.url, "Creating [testing] Postgres pool with config");
-        sqlx::migrate!().run(&pool).await?;
-        Ok(Self { pool })
-    }
-
-    pub fn into_shared(self) -> PersistentRepoShared {
-        Arc::new(Box::new(self))
-    }
-
-    pub async fn ping(conn: &mut PersistentDbConn) -> Result<(), DbError> {
-        db_helpers::ping(conn).await
-    }
-}
-
-#[async_trait]
-impl PersistentRepoTrait for PostgresRepoTesting {
-    async fn get_conn(&self) -> crate::error::Result<PersistentDbConn> {
-        Ok(self.pool.acquire().await?)
-    }
+    async fn get_conn(&self) -> Result<PersistentDbConn, DbGetConnError>;
 }
 
 impl PostgresRepo {
+    /// Provides **unmigrated** connection to db
+    ///
+    /// Why **unmigrated**? - it requires path to `migrate` folder as literal
     #[instrument(level = "trace", ret)]
     pub async fn from_config(creds: PostgresDbCredentials) -> Result<Self, DbInitError> {
-        trace!("Creating PG (production) connection pool with creds: {:?}", creds);
+        trace!("Creating PG connection pool with creds: {:?}", creds);
         let pool = PgPool::connect(&creds.url)
             .await
             .map_err(|x| DbInitError::FailedToEstablishDbConn(x, creds.url.clone()))?;
         trace!(db_url = creds.url, "Creating Postgres pool with config");
-        sqlx::migrate!().run(&pool).await?;
         Ok(Self { pool })
     }
 
@@ -73,14 +42,14 @@ impl PostgresRepo {
         Arc::new(Box::new(self))
     }
 
-    pub async fn ping(conn: &mut PersistentDbConn) -> Result<(), DbError> {
+    pub async fn ping(conn: &mut PersistentDbConn) -> Result<(), DbGetConnError> {
         db_helpers::ping(conn).await
     }
 }
 
 #[async_trait]
 impl PersistentRepoTrait for PostgresRepo {
-    async fn get_conn(&self) -> crate::error::Result<PersistentDbConn> {
+    async fn get_conn(&self) -> Result<PersistentDbConn, DbGetConnError> {
         Ok(self.pool.acquire().await?)
     }
 }
@@ -88,10 +57,10 @@ impl PersistentRepoTrait for PostgresRepo {
 pub mod db_helpers {
     use sqlx::Connection;
 
-    use crate::{error::DbError, init::PersistentDbConn};
+    use super::*;
 
     #[inline]
-    pub async fn ping(conn: &mut PersistentDbConn) -> Result<(), DbError> {
+    pub async fn ping(conn: &mut PersistentDbConn) -> Result<(), DbGetConnError> {
         Ok(conn.ping().await?)
     }
 }
